uid test

+++ game/src/objects/unit/UnitFactory.h
@@ -14,7 +14,7 @@ public:
 	UnitFactory();
 	~UnitFactory();
 
-	std::vector<Unit*>& create(unsigned int number, int id, Urho3D::Vector2& center, int playerId, int level);
+	std::vector<Unit*>& create(unsigned int number, short id, Urho3D::Vector2& center, short playerId, short level);
 	std::vector<Unit*>& load(dbload_unit* unit);
 private:
 	std::vector<Unit*> units;
